#28 Test07-runoff_regrid script updated

+++ t/Test07-runoff_regrid.sh
@@ -28,8 +28,9 @@
 		--output_mosaic ocean_mosaic.nc \
 		--output_topog topog.nc \
 		--output_file runoff.cm2m.nc
-  [ "$status" -eq 0 ]
-
+  if [ "$status" -eq 0 ]
+  then
   cd ..
   rm -rf Test07
+  fi
 }
